backend/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
gunicorn
faker
rich
python-dotenv
